Move OpenAPIProvider into the openapi plugin package
Aligns openapi with the pattern set by tool_search / code_mode /
prompts_as_tools / resources_as_tools: provider.py, components.py,
and routing.py move to fastmcp.server.plugins.openapi.*, with shims
at the old path for backcompat.

The shim's package __init__ stays silent so `from fastmcp.server.providers.openapi
import OpenAPIProvider` keeps working without spraying warnings on
every `import fastmcp`. Direct imports from leaf submodules
(.provider / .components / .routing) emit FastMCPDeprecationWarning
pointing at the new location.

Also moves the openapi test suite to tests/server/plugins/openapi/
and updates its imports to the new canonical path.

Part of FMCP-29.
