+++ tests/server/plugins/openapi/test_performance_comparison.py
@@ -7,7 +7,7 @@ import httpx
 import pytest
 
 from fastmcp import FastMCP
-from fastmcp.server.providers.openapi import OpenAPIProvider
+from fastmcp.server.plugins.openapi import OpenAPIProvider
 
 
 def create_openapi_server(
