+++ tests/server/plugins/openapi/test_server.py
@@ -5,8 +5,8 @@ import pytest
 
 from fastmcp import FastMCP
 from fastmcp.client import Client
-from fastmcp.server.providers.openapi import OpenAPIProvider
-from fastmcp.server.providers.openapi.provider import DEFAULT_TIMEOUT
+from fastmcp.server.plugins.openapi import OpenAPIProvider
+from fastmcp.server.plugins.openapi.provider import DEFAULT_TIMEOUT
 
 
 class TestOpenAPIProviderServerVariables:
